refactor(ui): split chat markdown renderer into cohesive modules
